Health update
Made Health float instead of int
Health of Enemies now Show above them as 3D Text

+++ Assets/Scripts/v1 Scripts/3 - Intermediate pt1/Interfaces/IDamagable.cs
@@ -4,11 +4,11 @@ using UnityEngine;
 
 public interface IDamagable
 {
-    int Health { get; set; }
+    float Health { get; set; }
 
     Vector3 RetPos();
 
     Color RetColor();
 
-    void Damage(int dmgAmount);
+    void Damage(float dmgAmount);
 }
